server/requirements/dev.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autopep8==2.3.2
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
carehare==1.0.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.3.0
channels-rabbitmq==4.0.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
coverage==7.10.1
cryptography==45.0.5
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-elasticsearch-dsl==9.0
django-filter==25.1
django-rest-passwordreset==1.5.0
django-silk==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docutils==0.22
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
elastic-transport==8.17.1
elasticsearch==9.0.2
factory_boy==3.3.3
fake-useragent==2.2.0
Faker==37.4.2
firebase_admin==7.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gprof2dot==2025.4.14
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
msgpack==1.1.1
outcome==1.3.0.post0
packaging==25.0
pamqp==3.3.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
selenium==4.34.2
sentry-sdk==2.33.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stripe==12.3.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
